feat: fixes and feats

+++ Assets/Scripts/Upgrade/IUpgradable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IUpgradable
 {
-    public float GetUpgradeTime();
+    public int GetUpgradeTime();
 }
